+++ test/integration/static-a-plain-filenames-test.js
@@ -13,14 +13,20 @@ let shared = join(tmp, 'node_modules', '@architect', 'shared')
 let origRegion = process.env.AWS_REGION
 let origCwd = process.cwd()
 
+let resetEnv = () => {
+  delete process.env.AWS_REGION
+  delete process.env.NODE_ENV
+  delete process.env.ARC_STATIC_FOLDER
+  delete process.env.ARC_STATIC_BUCKET
+}
+
 test('Set up mocked files', t=> {
   t.plan(2)
   mkdir(shared)
   fs.copyFileSync(join(mock, 'mock-arc'), join(shared, '.arc'))
   fs.copyFileSync(join(mock, 'mock-arc'), join(tmp, '.arc'))
-  fs.copyFileSync(join(mock, 'mock-static'), join(shared, 'static.json'))
-  t.ok(exists(join(shared, '.arc')), 'Mock .arc file ready')
-  t.ok(exists(join(shared, 'static.json')), 'Mock static.json file ready')
+  t.ok(exists(join(shared, '.arc')), 'Mock .arc (shared) file ready')
+  t.ok(exists(join(tmp, '.arc')), 'Mock .arc (root) file ready')
   process.chdir(tmp)
   // eslint-disable-next-line
   arc = require('../..') // module globally inspects arc file so need to require after chdir
@@ -38,13 +44,13 @@ test('Local URL tests', t=> {
   process.env.NODE_ENV = 'staging'
   t.notEqual(arc.static('index.html'), '/_static/index.html', 'Basic local static path not used in staging')
 
-  delete process.env.NODE_ENV
+  delete process.env.NODE_ENV // Run it "locally"
   process.env.ARC_STATIC_FOLDER = 'foo'
   t.equal(arc.static('index.html'), '/_static/index.html', 'Basic local static path unaffected by ARC_STATIC_FOLDER env var')
-  delete process.env.ARC_STATIC_FOLDER
+  resetEnv()
 })
 
-test('Staging and production URL tests', t=> {
+test('Staging and production URL tests (fingerprint disabled by lack of @architect/shared/static.json)', t=> {
   t.plan(5)
   process.env.AWS_REGION = 'us-west-1'
   process.env.NODE_ENV = 'production'
@@ -59,12 +65,31 @@ test('Staging and production URL tests', t=> {
 
   process.env.ARC_STATIC_FOLDER = 'a-folder'
   t.equals(arc.static('index.html'), 'https://a-totally-different-bucket.s3.us-west-1.amazonaws.com/a-folder/index.html', 'ARC_STATIC_FOLDER env var populates and matches')
+  resetEnv()
+})
+
+test('Staging and production URL tests (fingerprint inferred by @architect/shared/static.json)', t=> {
+  t.plan(6)
+  process.env.AWS_REGION = 'us-west-1'
+  process.env.NODE_ENV = 'production'
+  fs.copyFileSync(join(mock, 'mock-static'), join(shared, 'static.json'))
+  t.ok(exists(join(shared, 'static.json')), 'Mock static.json file ready')
+  t.equals(arc.static('index.html'), 'https://a-production-bucket.s3.us-west-1.amazonaws.com/index-1e25d663f6.html', 'Production URL matches')
+  t.equals(arc.http.helpers.static('index.html'), 'https://a-production-bucket.s3.us-west-1.amazonaws.com/index-1e25d663f6.html', 'Production URL matches (legacy)')
+
+  process.env.NODE_ENV = 'staging'
+  t.equals(arc.static('index.html'), 'https://a-staging-bucket.s3.us-west-1.amazonaws.com/index-1e25d663f6.html', 'Staging URL matches')
+
+  process.env.ARC_STATIC_BUCKET = 'a-totally-different-bucket'
+  t.equals(arc.static('index.html'), 'https://a-totally-different-bucket.s3.us-west-1.amazonaws.com/index-1e25d663f6.html', 'ARC_STATIC_BUCKET env var populates and matches')
+
+  process.env.ARC_STATIC_FOLDER = 'a-folder'
+  t.equals(arc.static('index.html'), 'https://a-totally-different-bucket.s3.us-west-1.amazonaws.com/a-folder/index-1e25d663f6.html', 'ARC_STATIC_FOLDER env var populates and matches')
+  resetEnv()
 })
 
 test('Clean up env', t=> {
   t.plan(1)
-  delete process.env.ARC_STATIC_BUCKET
-  delete process.env.ARC_STATIC_FOLDER
   process.env.AWS_REGION = origRegion
   process.env.NODE_ENV = 'testing'
   process.chdir(origCwd)